+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/blocks/LunchBoxBlock.java
@@ -1,0 +1,10 @@
+package net.mehvahdjukaar.supplementaries.common.block.blocks;
+
+import net.minecraft.world.level.block.Block;
+
+public class LunchBoxBlock extends Block {
+
+    public LunchBoxBlock(Properties properties) {
+        super(properties);
+    }
+}