+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/events/overrides/SuppAdditionalPlacement.java
@@ -3,6 +3,7 @@ package net.mehvahdjukaar.supplementaries.common.events.overrides;
 import net.mehvahdjukaar.moonlight.api.item.additional_placements.AdditionalItemPlacement;
 import net.minecraft.world.level.block.Block;
 
+//has tooltip. thats it
 public class SuppAdditionalPlacement extends AdditionalItemPlacement {
 
     public SuppAdditionalPlacement(Block placeable) {