+++ forge/src/main/java/net/mehvahdjukaar/supplementaries/client/renderers/forge/CannonChargeOverlayImpl.java
@@ -11,7 +11,6 @@ public class CannonChargeOverlayImpl extends CannonChargeOverlay implements IGui
     public CannonChargeOverlayImpl() {
         super(Minecraft.getInstance(), Minecraft.getInstance().getItemRenderer());
     }
-
     @Override
     public void render(ForgeGui forgeGui, GuiGraphics graphics, float f, int i, int j) {
         renderBar(graphics, f, i, j);